moved base model tests to seperate file, rewrote event test
