+++ service/tests/tests.py
@@ -1,8 +1,6 @@
 from django.test import TestCase
 from django.urls import resolve
 from service.views import dashboard
-from service.models import Event, Part, Container
-import datetime
 
 # Create your tests here.
 
@@ -16,75 +14,3 @@ class IndexViewTest(TestCase):
     def test_index_view_uses_index_template(self):
         response = self.client.get('/service/')
         self.assertTemplateUsed(response, 'service/dashboard.html')
-
-
-class EventModelTest(TestCase):
-
-    def test_saving_and_retrieving_events(self):
-
-        now = datetime.datetime.now()
-        today = now.strftime("%Y-%m-%d")
-
-        first_container = Container()
-        first_container.name = 'Tardis'
-        first_container.save()
-
-        first_event = Event()
-        first_event.title = "The first event"
-        first_event.description = "This is a test of the first event description"
-        first_event.action = "Install"
-        first_event.container = first_container
-        first_event.event_date = "05/10/2019"
-        first_event.save()
-
-        saved_events = Event.objects.all()
-        self.assertEqual(saved_events.count(), 1)
-
-        first_saved_event = saved_events[0]
-        self.assertEqual(first_saved_event.title, 'The first event')
-        self.assertEqual(first_saved_event.description, "This is a test of the first event description")
-        self.assertEqual(first_saved_event.action, 'Install')
-        self.assertEqual(first_saved_event.container, first_container)
-        self.assertEqual(first_saved_event.event_date, "2019-05-10")
-        self.assertEqual(first_saved_event.rec_date, today)
-
-
-class PartModelTest(TestCase):
-
-    def test_creating_and_retrieving_parts(self):
-
-        first_part = Part()
-        first_part.name = "Flux Capacitor"
-
-        first_part.save()
-
-        saved_parts = Part.objects.all()
-        self.assertEqual(saved_parts.count(), 1)
-
-        first_saved_part = saved_parts[0]
-        self.assertEqual(first_saved_part.name, 'Flux Capacitor')
-
-
-class ContainerModelTest(TestCase):
-    
-    def test_creating_and_retrieving_containers(self):
-        
-        first_container = Container()
-        first_container.name = 'Tardis'
-        first_container.save()
-
-        second_container = Container()
-        second_container.name = "England"
-        second_container.subcontainer = first_container
-        second_container.save()
-        
-        saved_containers = Container.objects.all()
-        self.assertEqual(saved_containers.count(), 2)
-        
-        first_saved_container = saved_containers[0]
-        self.assertEqual(first_saved_container.name, 'Tardis')
-
-        second_saved_container = saved_containers[1]
-        self.assertEqual(second_saved_container.name, 'England')
-        self.assertEqual(second_saved_container.subcontainer, first_container)
-        self.assertEqual(second_saved_container.subcontainer, first_saved_container)
